+++ PortfolioWeb_MarianoSuppo/Backend/src/main/java/com/portfolio/sma/Service/ImpPersonaService.java
@@ -28,14 +28,14 @@ public class ImpPersonaService implements IPersonaService{
     }
 
     @Override
-    public void deletePersona(Long id) {
+    public void deletePersona(int id) {
     
         ipersonaRepository.deleteById(id);
         
     }
 
     @Override
-    public Persona findPersona(Long id) {
+    public Persona findPersona(int id) {
     
         Persona persona = ipersonaRepository.findById(id).orElse(null);
         return persona;
